+++ x/tokenfactory/bindings/message_plugin.go
@@ -15,8 +15,8 @@ import (
 	bindingstypes "github.com/ChihuahuaChain/chihuahua/x/tokenfactory/bindings/types"
 	tokenfactorykeeper "github.com/ChihuahuaChain/chihuahua/x/tokenfactory/keeper"
 	tokenfactorytypes "github.com/ChihuahuaChain/chihuahua/x/tokenfactory/types"
-	liquiditykeeper "github.com/Victor118/liquidity/x/liquidity/keeper"
-	liquiditytypes "github.com/Victor118/liquidity/x/liquidity/types"
+	liquiditykeeper "github.com/ChihuahuaChain/liquidity/x/liquidity/keeper"
+	liquiditytypes "github.com/ChihuahuaChain/liquidity/x/liquidity/types"
 )
 
 // CustomMessageDecorator returns decorator for custom CosmWasm bindings messages